depl/ml-project/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
ultralytics>=8.0.0
Pillow>=10.0.0
pytest>=8.0.0
httpx>=0.26.0
